pyproject.toml
[project]
name = "seniorenradio"
version = "0.1.0"
description = "An easy-to-use and intuitive radio app for seniors to run on a Raspberry Pi"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0",
    "python-mpv>=1.0.0",
    "rpi-lgpio>=0.6; platform_system == 'Linux' and (platform_machine == 'armv7l' or platform_machine == 'aarch64')",
    "pydantic>=2.0",
]

[dependency-groups]
dev = [
    "pyrefly>=0.50.1",
    "ruff>=0.14.14",
]

[tool.ruff.lint]
extend-select = ["I", "B", "UP", "SIM", "N", "C4", "PTH", "ANN"]

[tool.pyrefly]
project-includes = ["src/**/*.py"]